+++ hallEffect.hpp
@@ -27,7 +27,7 @@ namespace Sensor {
         ~HallEffect();
 
         void tick();
-        double read();
+        float read();
     };
 
     HallEffect::HallEffect() {
@@ -56,13 +56,13 @@ namespace Sensor {
         // currentTime += 1.0 / 1000.0;
     }
 
-    double HallEffect::read() {
+    float HallEffect::read() {
         unsigned long currentTime = millis();
         int timeDelta = currentTime - prevTime;
 
-        double time = (double)timeDelta / 1000.0;
+        float time = (float)timeDelta / 1000.0;
 
-        double rpm = ((double)ticks / (double)ticksPerRev) / (time / 60.0);
+        float rpm = ((float)ticks / (float)ticksPerRev) / (time / 60.0);
 
         prevTime = currentTime;
         ticks = 0;
